+++ CHANGES.md
@@ -3,9 +3,11 @@
 
 * Support for multiple Chef cookbook source directories
   when using the `vm config add cookbook` command.
-* Fix Support different handling of here-doc in Zsh/Bash. 
 * Adding support for Chef data-bags used together with 
   _chef-solo_.
+* [BUG] Support different handling of here-doc in Zsh/Bash. 
+* [BUG] Make sure to always alert users to add cookbooks,
+  before running, `vm config solo`
 
 # 1.0